Update linux.ts
Added  "Wants=network-online.target" to the file. This sets a "dependency" that the networking online starts before Homebridge. From "man systemd.unit -  It is a common pattern to include a unit name in both the After= and Requires/Wants= options, in which case the
unit listed will be started before the unit that is configured with these options". In my case my Raspberry Pi sometimes starts homebridge plugins before the network is fully up leading to gethostaddr errors.