Update es.json (Improves Spanish localization) (#1632)
Major improvements to the Spanish localization.

### :recycle: Current situation

The current Spanish localization of the Homebridge UI has room for
improvement. Some phrases and terms used are not the most accurate or
natural translations, which can lead to confusion for Spanish-speaking
users. For example, the term "triggered" was translated as "activado",
which is incorrect.

### :bulb: Proposed solution

This pull request introduces major improvements to the Spanish
localization of the Homebridge UI. The changes include more accurate
translations of technical terms, better grammar, and more natural
sounding phrases. These changes do not affect the internal structure of
the project, but they significantly improve the user experience for
Spanish-speaking users.

### :gear: Release Notes

From a user's point of view, Spanish-speaking users will notice a
significant improvement in the translations. The language will sound
more natural and less like a direct translation from English. This will
make the Homebridge UI more accessible and user-friendly for
Spanish-speaking users. There are no breaking changes introduced with
this pull request.